Доработки

+++ src/main/java/org/example/Task1_9.java
@@ -1,6 +1,5 @@
 package org.example;
 
-import javax.swing.*;
 
 public class Task1_9 {
     public static void main(String[] args) {
@@ -17,7 +16,7 @@ public class Task1_9 {
         int sum = 0;
 
         for(int i = 0; i < mas.length; ++i) {
-            sum = (int)((double)sum + Math.pow((double)mas[i], 3.0));
+            sum = (int)(sum + Math.pow(mas[i], 3.0));
         }
 
         return sum;
